+++ android/src/main/kotlin/com/hypertrack/sdk/flutter/HyperTrackPlugin.kt
@@ -1,14 +1,11 @@
 package com.hypertrack.sdk.flutter
 
-
-import android.content.Context
-import android.location.Location
-import android.util.Log
-import androidx.annotation.NonNull
-import com.hypertrack.sdk.GeotagResult
-import com.hypertrack.sdk.HyperTrack
-import com.hypertrack.sdk.TrackingError
-import com.hypertrack.sdk.TrackingStateObserver
+import com.hypertrack.sdk.*
+import com.hypertrack.sdk.flutter.common.*
+import com.hypertrack.sdk.flutter.common.Result
+import com.hypertrack.sdk.flutter.common.Serialization.serializeIsAvailable
+import com.hypertrack.sdk.flutter.common.Serialization.serializeIsTracking
+import com.hypertrack.sdk.flutter.common.Success
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.EventChannel
@@ -17,219 +14,232 @@ import io.flutter.plugin.common.EventChannel.StreamHandler
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
-import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry.Registrar
+import java.lang.NullPointerException
 import java.util.*
 
-/** HyperTrackPlugin */
-public class HyperTrackPlugin(): FlutterPlugin, MethodCallHandler, StreamHandler {
+public class HyperTrackPlugin : FlutterPlugin, MethodCallHandler {
 
-  private var applicationContext : Context? = null
-  private var methodChannel : MethodChannel? = null
-  private var eventChannel : EventChannel? = null
+    // receives method calls from the plugin API
+    private var methodChannel: MethodChannel? = null
 
-  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    Log.i(TAG, "onAttachedToEngine")
-    val context = flutterPluginBinding.applicationContext
-    val binaryMessenger = flutterPluginBinding.binaryMessenger
-    onAttachedToEngine(context, binaryMessenger)
-  }
+    // send events from the SDK to plugin clients
+    private var trackingStateEventChannel: EventChannel? = null
+    private var availabilityEventChannel: EventChannel? = null
+    private var errorEventChannel: EventChannel? = null
 
-  private fun onAttachedToEngine(context: Context, messenger: BinaryMessenger) {
-    applicationContext = context
-    methodChannel = MethodChannel(messenger, HYPERTRACK_SDK_METHOD_CHANNEL)
-    methodChannel?.setMethodCallHandler(this)
-    eventChannel = EventChannel(messenger, HYPERTRACK_SDK_STATE_CHANNEL)
-    eventChannel?.setStreamHandler(this)
-  }
+    private var trackingStateListener: TrackingStateObserver.OnTrackingStateChangeListener? = null
+    private var errorChannelTrackingStateListener: TrackingStateObserver.OnTrackingStateChangeListener? =
+        null
+    private var availabilityListener: AvailabilityStateObserver.OnAvailabilityStateChangeListener? =
+        null
 
-  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-    applicationContext = null
-    methodChannel?.setMethodCallHandler(null)
-    methodChannel = null
-    eventChannel?.setStreamHandler(null)
-    eventChannel = null
-  }
-
-  companion object {
-    private const val TAG = "SdkPlugin"
-    private const val HYPERTRACK_SDK_METHOD_CHANNEL = "sdk.hypertrack.com/handle"
-    private const val HYPERTRACK_SDK_STATE_CHANNEL = "sdk.hypertrack.com/trackingState"
-
-    @JvmStatic
-    fun registerWith(registrar: Registrar) {
-      Log.i(TAG, "registerWith")
-
-      registrar.activity()?.applicationContext?.let { context -> 
-        val messenger = registrar.messenger()
-        HyperTrackPlugin().onAttachedToEngine(context, messenger)
-      } ?: run {
-        Log.i(TAG, "failedRegister")
-      }
-    }
-  }  
-  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
-    if (call.method == "initialize") {
-      initialize(call.arguments(), result)
-      return
-    } else if (call.method == "enableDebugLogging") {
-      enableDebugLogging()
-      return
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+        val messenger = flutterPluginBinding.binaryMessenger
+        methodChannel = MethodChannel(messenger, METHOD_CHANNEL_NAME)
+        methodChannel?.setMethodCallHandler(this)
+        initEventChannels(messenger)
     }
 
-    val sdk = sdkInstance
-    if (sdk == null) {
-      result.error("NOT_INITIALIZED", "Internal Error: onMethodCall(${call.method}) - sdkInstance is null", null)
-      return
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        methodChannel?.setMethodCallHandler(null)
+        methodChannel = null
+        trackingStateEventChannel?.setStreamHandler(null)
+        trackingStateEventChannel = null
+        availabilityEventChannel?.setStreamHandler(null)
+        availabilityEventChannel = null
     }
 
-    when (call.method) {
-      "getDeviceId" -> result.success(sdk.deviceID)
-      "isRunning" -> result.success(sdk.isRunning)
-      "start" -> start(result, sdk)
-      "stop" -> stop(result, sdk)
-      "addGeotag" -> call.arguments<Map<String, Any>>()?.let { arguments -> addGeotag(arguments, result, sdk) } ?: result.error("INVALID_ARGS", "Internal Error: onMethodCall(${call.method}) - arguments is null", null)
-      "allowMockLocations" -> allowMockLocations(result, sdk)
-      "setDeviceName" -> call.arguments<String>()?.let { arguments -> setDeviceName(arguments, result, sdk) } ?: result.error("INVALID_ARGS", "Internal Error: onMethodCall(${call.method}) - arguments is null", null)
-      "setDeviceMetadata" -> call.arguments<Map<String, Any>?>()?.let { arguments -> setDeviceMetadata(arguments, result, sdk) } ?: result.error("INVALID_ARGS", "Internal Error: onMethodCall(${call.method}) - arguments is null", null)
-      "syncDeviceSettings" -> syncDeviceSettings(result, sdk)
-      else -> result.notImplemented()
-    }
-  }
-
-  private var sdkInstance : HyperTrack? = null
-  private var stateListener : TrackingStateObserver.OnTrackingStateChangeListener? = null
-
-  private fun initialize(publishableKey : String?, result: Result) {
-    Log.d(TAG, "getInstance for key $publishableKey")
-    if (publishableKey == null) return
-
-    try {
-      sdkInstance = HyperTrack.getInstance(publishableKey)
-      result.success(null)
-      return
-    } catch (e: Exception) {
-      result.error("INIT_ERROR", e.message, e)
-    }
-  }
-
-  private fun enableDebugLogging() {
-    Log.d(TAG, "enableDebugLogging called")
-
-    HyperTrack.enableDebugLogging()
-  }
-
-  private fun start(result: Result, sdk : HyperTrack) {
-    Log.d(TAG, "start")
-
-    val listener = object : TrackingStateObserver.OnTrackingStateChangeListener {
-      override fun onTrackingStart() {
-        Log.d(TAG, "onTrackingStart")
-        result.success(null)
-        sdk.removeTrackingListener(this)
-      }
-
-      override fun onError(p0: TrackingError?) {
-        Log.d(TAG, "onError " + p0?.message)
-        result.error("Start failed", p0?.message?:"", null)
-        sdk.removeTrackingListener(this)
-      }
-
-      override fun onTrackingStop() { /* NOOP */ }
+    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        invokeSdkMethod(call).sendAsFlutterResult(call, result)
     }
 
-    sdk.addTrackingListener(listener).start()
-
-  }
-
-  private fun stop(result: Result, sdk : HyperTrack) {
-    Log.d(TAG, "stop")
-    sdk.stop()
-    result.success(null)
-  }
-
-  private fun addGeotag(options : Map<String, Any>, result: Result, sdk : HyperTrack) {
-    val data = options["data"] as Map<String, java.io.Serializable>?
-    data?.let {
-      val expectedLocation: Location? = (options["expectedLocation"] as Map<String, Any>?)?.let { params ->
-        val lat = params["latitude"] as Double
-        val lng = params["longitude"] as Double
-        Location("any").apply{
-          latitude = lat
-          longitude = lng
+    private fun invokeSdkMethod(
+        call: MethodCall
+    ): Result<*> {
+        val method = SdkMethod
+            .values()
+            .firstOrNull { it.name == call.method }
+            ?: run {
+                return Success(NotImplemented)
+            }
+        return when (method) {
+            SdkMethod.initialize -> {
+                withArgs<Unit>(call) { args ->
+                    HyperTrackSdkWrapper
+                        .initializeSdk(args)
+                        .mapSuccess {
+                            Unit
+                        }
+                }
+            }
+            SdkMethod.getDeviceID -> {
+                HyperTrackSdkWrapper.getDeviceId()
+            }
+            SdkMethod.isTracking -> {
+                HyperTrackSdkWrapper.isTracking()
+            }
+            SdkMethod.isAvailable -> {
+                HyperTrackSdkWrapper.isAvailable()
+            }
+            SdkMethod.setAvailability -> {
+                withArgs<Unit>(call) { args ->
+                    HyperTrackSdkWrapper.setAvailability(args)
+                }
+            }
+            SdkMethod.getLocation -> {
+                HyperTrackSdkWrapper.getLocation()
+            }
+            SdkMethod.startTracking -> {
+                HyperTrackSdkWrapper.startTracking()
+            }
+            SdkMethod.stopTracking -> {
+                HyperTrackSdkWrapper.stopTracking()
+            }
+            SdkMethod.addGeotag -> {
+                withArgs<Map<String, Any?>>(call) { args ->
+                    HyperTrackSdkWrapper.addGeotag(args)
+                }
+            }
+            SdkMethod.setName -> {
+                withArgs<Unit>(call) { args ->
+                    HyperTrackSdkWrapper.setName(args)
+                }
+            }
+            SdkMethod.setMetadata -> {
+                withArgs<Unit>(call) { args ->
+                    HyperTrackSdkWrapper.setMetadata(args)
+                }
+            }
+            SdkMethod.sync -> {
+                HyperTrackSdkWrapper.sync()
+            }
         }
-      }
-      when (val geotagResult = sdk.addGeotag(data, expectedLocation)) {
-          is GeotagResult.SuccessWithDeviation -> result.success("""{"result": "success", "distance": ${geotagResult.deviationDistance}}""")
-          is GeotagResult.Success -> result.success("""{"result": "success"}""")
-          is GeotagResult.Error -> result.success("""{"result": "error", "reason": "${geotagResult.reason}"}""")
-        }
-    } ?: result.error("GEOTAG_ERROR", "No geotag data provided", null)
-  }
-
-  private fun allowMockLocations(result : Result, sdk : HyperTrack) {
-    Log.d(TAG, "allowMockLocations called")
-
-    sdk.allowMockLocations()
-    result.success(null)
-  }
-
-  private fun setDeviceName(name : String, result : Result, sdk : HyperTrack) {
-    Log.d(TAG, "setDeviceName called with name $name")
-
-    sdk.setDeviceName(name)
-    result.success(null)
-  }
-
-  private fun setDeviceMetadata(data : Map<String, Any>, result: Result, sdk : HyperTrack) {
-    Log.d(TAG, "setDeviceMetadata called with data $data")
-
-    sdk.setDeviceMetadata(data)
-    result.success(null)
-  }
-
-  private fun syncDeviceSettings(result : Result, sdk : HyperTrack) {
-    Log.d(TAG, "syncDeviceSettings called")
-
-    sdk.syncDeviceSettings()
-    result.success(null)
-  }
-
-  override fun onListen(arguments: Any?, events: EventSink?) {
-    val sdk = sdkInstance
-    if (sdk == null) {
-      events?.error("INVALID_STATE", "Cannot create stream before sdk init", null)
-      return
     }
 
-    stateListener = object : TrackingStateObserver.OnTrackingStateChangeListener {
-      override fun onTrackingStart() {
-        events?.success("start")
-      }
-
-      override fun onError(p0: TrackingError?) {
-        when (p0?.code) {
-          TrackingError.INVALID_PUBLISHABLE_KEY_ERROR -> events?.success("publishable_key_error")
-          TrackingError.PERMISSION_DENIED_ERROR -> events?.success("permissions_denied")
-          TrackingError.AUTHORIZATION_ERROR -> events?.success("auth_error")
-          TrackingError.GPS_PROVIDER_DISABLED_ERROR -> events?.success("gps_disabled")
-          TrackingError.UNKNOWN_NETWORK_ERROR -> events?.success("network_error")
-          else -> events?.success(p0?.message?:"unknown error")
-        }
-      }
-
-      override fun onTrackingStop() {
-        events?.success("stop")
-      }
+    private fun <N> withArgs(call: MethodCall, block: (Map<String, Any?>) -> Result<N>): Result<N> {
+        return call
+            .arguments<Map<String, Any?>>()
+            ?.let { block.invoke(it) }
+            ?: Failure(NullPointerException(call.method))
     }
 
-    sdk.addTrackingListener(stateListener)
+    private fun initEventChannels(messenger: BinaryMessenger) {
+        trackingStateEventChannel = EventChannel(messenger, TRACKING_STATE_EVENT_CHANNEL_NAME)
+        trackingStateEventChannel?.setStreamHandler(object : StreamHandler {
+            override fun onListen(arguments: Any?, events: EventSink) {
+                Result
+                    .tryAsResult {
+                        val sdk = HyperTrackSdkWrapper.sdkInstance
+                        trackingStateListener =
+                            object : TrackingStateObserver.OnTrackingStateChangeListener {
+                                override fun onTrackingStart() {
+                                    events.success(serializeIsTracking(true))
+                                }
 
-  }
+                                override fun onTrackingStop() {
+                                    events.success(serializeIsTracking(false))
+                                }
 
-  override fun onCancel(arguments: Any?) {
-    sdkInstance?.removeTrackingListener(stateListener)
-    stateListener = null
-  }
+                                override fun onError(p0: TrackingError) {
+                                    // ignored, errors are handled by errorEventChannel
+                                }
+                            }
+                        sdk.addTrackingListener(trackingStateListener)
+                        events.success(serializeIsTracking(sdk.isTracking))
+                    }
+                    .crashAppIfError()
+            }
+
+            override fun onCancel(arguments: Any?) {
+                HyperTrackSdkWrapper.sdkInstance
+                    .let { sdk ->
+                        sdk.removeTrackingListener(trackingStateListener)
+                        trackingStateListener = null
+                    }
+            }
+        })
+
+        errorEventChannel = EventChannel(messenger, ERROR_EVENT_CHANNEL_NAME)
+        errorEventChannel?.setStreamHandler(object : StreamHandler {
+            override fun onListen(arguments: Any?, events: EventSink) {
+                Result
+                    .tryAsResult {
+                        val sdk = HyperTrackSdkWrapper.sdkInstance
+                        errorChannelTrackingStateListener =
+                            object : TrackingStateObserver.OnTrackingStateChangeListener {
+                                override fun onTrackingStart() {
+                                    // ignored, trackingState is handled by trackingStateEventChannel
+                                }
+
+                                override fun onTrackingStop() {
+                                    // ignored, trackingState is handled by trackingStateEventChannel
+                                }
+
+                                override fun onError(error: TrackingError) {
+                                    events.success(HyperTrackSdkWrapper.getErrors(error))
+                                }
+                            }
+                        sdk.addTrackingListener(errorChannelTrackingStateListener)
+                        events.success(HyperTrackSdkWrapper.getInitialErrors())
+                    }
+                    .crashAppIfError()
+            }
+
+            override fun onCancel(arguments: Any?) {
+                HyperTrackSdkWrapper.sdkInstance
+                    .let { sdk ->
+                        sdk.removeTrackingListener(errorChannelTrackingStateListener)
+                        errorChannelTrackingStateListener = null
+                    }
+            }
+        })
+
+        availabilityEventChannel = EventChannel(messenger, AVAILABILTY_EVENT_CHANNEL_NAME)
+        availabilityEventChannel?.setStreamHandler(
+            object : StreamHandler {
+                override fun onListen(arguments: Any?, events: EventSink) {
+                    Result.tryAsResult {
+                        val sdk = HyperTrackSdkWrapper.sdkInstance
+                        availabilityListener =
+                            object :
+                                AvailabilityStateObserver.OnAvailabilityStateChangeListener {
+                                override fun onError(p0: AvailabilityError) {
+                                    // ignored, errors are handled by errorEventChannel
+                                }
+
+                                override fun onAvailable() {
+                                    events.success(serializeIsAvailable(true))
+                                }
+
+                                override fun onUnavailable() {
+                                    events.success(serializeIsAvailable(false))
+                                }
+                            }
+                        sdk.addAvailabilityListener(availabilityListener)
+                        events.success(
+                            serializeIsAvailable(sdk.availability.equals(Availability.AVAILABLE))
+                        )
+                    }
+                        .crashAppIfError()
+                }
+
+                override fun onCancel(arguments: Any?) {
+                    HyperTrackSdkWrapper.sdkInstance
+                        .let { sdk ->
+                            sdk.removeAvailabilityListener(availabilityListener)
+                            availabilityListener = null
+                        }
+                }
+            }
+        )
+    }
+
+    companion object {
+        private const val METHOD_CHANNEL_NAME = "sdk.hypertrack.com/methods"
+        private const val TRACKING_STATE_EVENT_CHANNEL_NAME = "sdk.hypertrack.com/tracking"
+        private const val ERROR_EVENT_CHANNEL_NAME = "sdk.hypertrack.com/errors"
+        private const val AVAILABILTY_EVENT_CHANNEL_NAME = "sdk.hypertrack.com/availability"
+
+        internal const val ERROR_CODE_METHOD_CALL = "method_call_error"
+    }
 }
-