+++ lib/data_types/location.dart
@@ -1,0 +1,36 @@
+class Location {
+  double latitude;
+  double longitude;
+
+  Location(this.latitude, this.longitude);
+
+  /// @nodoc
+  @override
+  String toString() {
+    return "{$latitude, $longitude}";
+  }
+
+  /// @nodoc
+  @override
+  dynamic noSuchMethod(Invocation invocation) {
+    return super.noSuchMethod(invocation);
+  }
+
+  /// @nodoc
+  @override
+  bool operator ==(Object other) {
+    return super == other;
+  }
+
+  /// @nodoc
+  @override
+  int get hashCode {
+    return super.hashCode;
+  }
+
+  /// @nodoc
+  @override
+  Type get runtimeType {
+    return super.runtimeType;
+  }
+}