+++ lib/src/sdk_method.dart
@@ -1,0 +1,15 @@
+// using Swift naming convention
+enum SdkMethod {
+  initialize,
+  getDeviceID,
+  getLocation,
+  startTracking,
+  stopTracking,
+  setAvailability,
+  setName,
+  setMetadata,
+  isTracking,
+  isAvailable,
+  addGeotag,
+  sync
+}