+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.0.0] - 2023-02-17
+
+### Changed
+- Updated HyperTrack iOS SDK to 4.14.0
+- Updated HyperTrack Android SDK to 6.4.0
+- `syncDeviceSettings()` renamed to `sync()`
+- `setDeviceName()` renamed to `setName()`
+- `setDeviceMetadata()` renamed to `setMetadata()`
+- `start()` renamed to `startTracking()`
+- `stop()` renamed to `stopTracking()`
+
+### Added
+- `initialize()` configuration params for
+  - Debug logging
+  - Background location permissions request for Android
+  - Mock locations
+- `onAvailabilityChanged` stream
+- `onError` stream
+- Location result for `addGeotag`
+
+### Removed
+- `getLatestLocation()`
+- `allowMockLocations()` (use `initialize()` param `allowMockLocations` istead)
+- `enableDebugLogging()` (use `initialize()` param `loggingEnabled` istead)
+- `getRunnigStatus()`
+
 ## [0.4.3] - 2022-09-16
 #### Changed
 - Android SDK updated to 6.3.0